# swagger.yaml
openapi: 3.0.0
paths:
...
  /parking-record/entry:
    post:
      operationId: ParkingRecordsController_createParkingEntryRecord
      summary: Create parking entry record
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingRecordRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingRecordEntryResponseDto'
        '400':
          description: >-
            Bad Request - This endpoint can return multiple responses based on
            bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                        - The parking not assigned to the user
                        - Parking is full
                        - >-
                          Unable to Register Entry, vehicle already in a parking
                          lot
                      description: List of errors
                  error:
                    type: string
                    example: >-
                      BadRequestException, ParkingNotAssignedException,
                      ParkingIsFullException, ParkingNotExistsException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.410Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.410Z'
                    description: Date of the error
      tags: &ref_3
        - Parking Records
      security: &ref_4
        - access-token: []
  /parking-record/exit:
    post:
      operationId: ParkingRecordsController_createParkingExitRecord
      summary: Create parking exit record
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingRecordRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingRecordExitResponseDto'
        '400':
          description: >-
            Bad Request - This endpoint can return multiple responses based on
            bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                        - The parking not assigned to the user
                        - Vehicle not found
                        - >-
                          Unable to Register Exit, no license plate found in any
                          parking lot.
                        - >-
                          Unable to Register Exit, vehicle license plate in
                          another parking lot.
                        - >-
                          Unable to Register Exit, vehicle not found in a
                          parking lot.
                      description: List of errors
                  error:
                    type: string
                    example: >-
                      BadRequestException, ParkingNotAssignedException,
                      VehicleNotFoundException, ParkingNotExistsException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.411Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.411Z'
                    description: Date of the error
      tags: *ref_3
      security: *ref_4
  /parking-record/parked-vehicles/parkingId/{parkingId}:
    get:
      operationId: ParkingRecordsController_getParkingRecordsWithExitDateNullByParkingId
      summary: Get the vehicles currently parked in a specific parking lot
      parameters:
        - name: parkingId
          required: true
          in: path
          description: Parking id
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingRecordEntryResponseDto'
        '400':
          description: >-
            Bad Request - This endpoint can return multiple responses based on
            bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                        - The parking not assigned to the user
                      description: List of errors
                  error:
                    type: string
                    example: ParkingNotExistsException, ParkingNotAssignedException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.411Z'
                    description: Date of the error
      tags: *ref_3
      security: *ref_4
  /parking-record/parked-vehicles/most-registered:
    get:
      operationId: ParkingRecordsController_getMostRegisteredVehiclesAtAllParking
      summary: >-
        Find the top 10 vehicles that have been registered the most times in all
        parking lots and how many times they have been registered
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponseDto'
      tags: *ref_3
      security: *ref_4
  /parking-record/parked-vehicles/most-registered/parkingId/{parkingId}:
    get:
      operationId: >-
        ParkingRecordsController_getMostRegisteredVehiclesAtAllParkingByParkingId
      summary: >-
        Find the 10 vehicles that have been registered the most times in a
        specific parking lot and how many times they have been registered
      parameters:
        - name: parkingId
          required: true
          in: path
          description: Parking id
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponseDto'
        '400':
          description: >-
            Bad Request - This endpoint can return multiple responses based on
            bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                        - The parking not assigned to the user
                      description: List of errors
                  error:
                    type: string
                    example: ParkingNotExistsException, ParkingNotAssignedException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.411Z'
                    description: Date of the error
      tags: *ref_3
      security: *ref_4
  /parking-record/parked-vehicles/first-time/parkingId/{parkingId}:
    get:
      operationId: >-
        ParkingRecordsController_getParkingRecordsFirstTimeWithExitDateNullByParkingId
      summary: Obtain vehicles parked for the first time in that parking lot
      parameters:
        - name: parkingId
          required: true
          in: path
          description: Parking id
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingRecordEntryResponseDto'
        '400':
          description: >-
            Bad Request - This endpoint can return multiple responses based on
            bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Parking not exists
                        - The parking not assigned to the user
                      description: List of errors
                  error:
                    type: string
                    example: ParkingNotExistsException, ParkingNotAssignedException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.412Z'
                    description: Date of the error
      tags: *ref_3
      security: *ref_4
  /parking-record/parked-vehicles/matches/plate/{plate}:
    get:
      operationId: ParkingRecordsController_getParkingRecordsByVehiclePlateMatches
      summary: Obtain vehicles parked for the first time in all parking lots
      parameters:
        - name: plate
          required: true
          in: path
          description: Vehicle plate number
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingRecordEntryResponseDto'
      tags: *ref_3
      security: *ref_4
...
info:
  title: PARKING CRUD API
  description: >-
    The Rest API manages vehicle parking, allows partner creation for vehicle
    entry and exit, and uses JWT tokens with role-based access control (ADMIN
    and SOCIO) expiring every 6 hours, built with NestJS and PostgreSQL
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
  schemas:
...
    ParkingRecordRequestDto:
      type: object
      properties:
        plate:
          type: string
          description: Vehicle plate
          example: ABC123
          minLength: 6
          maxLength: 6
        parkingId:
          type: number
          description: Parking id
          example: 1
      required:
        - plate
        - parkingId
    ParkingRecordEntryResponseDto:
      type: object
      properties:
        id:
          type: number
          description: Record id
          example: 1
        plate:
          type: string
          description: Vehicle plate
          example: ABC123
        entryDate:
          format: date-time
          type: string
          description: Entry date
          example: '2021-10-15T00:00:00.000Z'
      required:
        - id
        - plate
        - entryDate
    ParkingRecordExitResponseDto:
      type: object
      properties:
        id:
          type: number
          description: Record id
          example: 1
        plate:
          type: string
          description: Vehicle plate
          example: ABC123
        entryDate:
          format: date-time
          type: string
          description: Entry date
          example: '2021-10-15T00:00:00.000Z'
        exitDate:
          format: date-time
          type: string
          description: Exit date
          example: '2021-10-15T04:00:00.000Z'
      required:
        - id
        - plate
        - entryDate
        - exitDate
    VehicleResponseDto:
      type: object
      properties:
        id:
          type: number
          description: Vehicle id
          example: 1
        plate:
          type: string
          description: Vehicle plate
          example: ABC123
        recordCount:
          type: number
          description: Record count
          example: 2
      required:
        - id
        - plate
        - recordCount
...